Run storefront tests on demo release (#804)
This PR runs the tests on demo release, then collects the results from
cypress and checks if they passed. If passed, posts a comment to PR on
cloud-deployments. If they didn't pass any, it checks if there are any
issues created for these tests on github. If there are no issues, it
creates them and in the comment to the release PR with links to issues
and to the cypress dashboard for the tester to check. If there are only
issues marked as "known bugs", we will have comment that tests with
issues marked as known bugs are failed, and PR can be merged